Player Params set to Float type

+++ trunk/service/src/main/java/com/fm/service/IPlayerGenerationStrategy.java
@@ -16,11 +16,11 @@ public interface IPlayerGenerationStrategy
 
    int getPotential(int age);
 
-   int getHighAbility();
+   float getHighAbility();
 
-   int getStandardAbility();
+   float getStandardAbility();
 
-   int getLowAbility();
+   float getLowAbility();
 
    Player getGoalkeeper(Team team);
